requirements.txt
requests>=2.27.1
PyYAML>=6.0
rich>=12.0.0
colorama>=0.4.4
importlib-metadata>=4.11.3
urllib3>=1.26.9
certifi>=2022.3.29
charset-normalizer>=2.0.12
idna>=3.3
jsonschema>=4.4.0
python-dateutil>=2.8.2
six>=1.16.0
markupsafe>=2.1.1
jinja2>=3.1.1
pygments>=2.12.0
concurrent-log-handler>=0.9.20 